perf: use multi-run averaging for microsecond-level timing accuracy

+++ assignment_03/src/gradient_descent.h
@@ -18,7 +18,7 @@ struct OptimizationResult {
     double final_fx = 0.0;
     int iterations_completed = 0;
     bool is_converged = false;
-    long long elapsed_time_microseconds = 0;
+    double elapsed_time_microseconds = 0.0;
 };
 
 class GradientDescentSolver {
